Begin separating tests into separate DBDataCreator commands
