+++ src/core/tasks/url/operators/agency_identification/dtos/suggestion.py
@@ -7,7 +7,7 @@ from src.core.enums import SuggestionType
 
 class URLAgencySuggestionInfo(BaseModel):
     url_id: int
-    suggestion_type: SuggestionType
+    suggestion_type: SuggestionType = SuggestionType.UNKNOWN
     pdap_agency_id: Optional[int] = None
     agency_name: Optional[str] = None
     state: Optional[str] = None
