+++ src/core/tasks/url/operators/agency_identification/queries/get_pending_urls_without_agency_suggestions.py
@@ -22,8 +22,8 @@ class GetPendingURLsWithoutAgencySuggestionsQueryBuilder(QueryBuilderBase):
             )
             .select_from(URL)
             .where(URL.outcome == URLStatus.PENDING.value)
-            .join(LinkBatchURL)
-            .join(Batch)
+            .outerjoin(LinkBatchURL)
+            .outerjoin(Batch)
         )
         statement = StatementComposer.exclude_urls_with_agency_suggestions(statement)
         statement = statement.limit(100)
