+++ src/core/tasks/url/operators/agency_identification/subtasks/ckan.py
@@ -1,12 +1,15 @@
-from typing import Optional
+from typing import Optional, __all__, final
+
+from typing_extensions import override
 
 from src.core.helpers import process_match_agency_response_to_suggestions
 from src.core.tasks.url.operators.agency_identification.dtos.suggestion import URLAgencySuggestionInfo
+from src.core.tasks.url.operators.agency_identification.subtasks.base import AgencyIdentificationSubtaskBase
 from src.external.pdap.client import PDAPClient
 from src.external.pdap.dtos.match_agency.response import MatchAgencyResponse
 
-
-class CKANAgencyIdentificationSubtask:
+@final
+class CKANAgencyIdentificationSubtask(AgencyIdentificationSubtaskBase):
 
     def __init__(
             self,
@@ -14,10 +17,11 @@ class CKANAgencyIdentificationSubtask:
     ):
         self.pdap_client = pdap_client
 
+    @override
     async def run(
             self,
             url_id: int,
-            collector_metadata: Optional[dict]
+            collector_metadata: dict | None = None
     ) -> list[URLAgencySuggestionInfo]:
         agency_name = collector_metadata["agency_name"]
         match_agency_response: MatchAgencyResponse = await self.pdap_client.match_agency(
