+++ tests/automated/integration/tasks/url/agency_identification/subtasks/test_auto_googler.py
@@ -1,0 +1,18 @@
+import pytest
+
+from src.core.enums import SuggestionType
+from src.core.tasks.url.operators.agency_identification.dtos.suggestion import URLAgencySuggestionInfo
+from src.core.tasks.url.operators.agency_identification.subtasks.auto_googler import \
+    AutoGooglerAgencyIdentificationSubtask
+from tests.helpers.data_creator.core import DBDataCreator
+
+
+@pytest.mark.asyncio
+async def test_auto_googler_subtask():
+    # Test that auto_googler subtask correctly adds URL to
+    # url_agency_suggestions with label 'Unknown'
+    subtask = AutoGooglerAgencyIdentificationSubtask()
+    results: list[URLAgencySuggestionInfo] = await subtask.run(url_id=1, collector_metadata={})
+    assert len(results) == 1
+    assert results[0].url_id == 1
+    assert results[0].suggestion_type == SuggestionType.UNKNOWN
