+++ tests/helpers/setup/final_review/core.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from src.api.endpoints.annotate.agency.post.dto import URLAgencyAnnotationPostInfo
-from src.core.enums import RecordType
+from src.core.enums import RecordType, SuggestedStatus
 from tests.helpers.data_creator.core import DBDataCreator
 from tests.helpers.setup.final_review.model import FinalReviewSetupInfo
 
@@ -46,7 +46,7 @@ async def setup_for_get_next_url_for_final_review(
     async def add_relevant_suggestion(relevant: bool):
         await db_data_creator.user_relevant_suggestion(
             url_id=url_mapping.url_id,
-            relevant=relevant
+            suggested_status=SuggestedStatus.RELEVANT if relevant else SuggestedStatus.NOT_RELEVANT
         )
 
     await db_data_creator.auto_relevant_suggestions(
